+++ R/getRecount.R
@@ -9,69 +9,90 @@
 #' @return Saves data to working directory.
 #'
 #' @export
-getRecount <- function(ID, characteristic = 1){
+getRecount <- function(ID, characteristic){
 
   catchwd <- getwd()
-
-  miSciTools::packageCheck("recount")
-  miSciTools::packageCheck("SummarizedExperiment")
-
-  ## Download data summarized at gene-level
+  packageCheck("recount")
+  packageCheck("SummarizedExperiment")
   url <- recount::download_study(ID)
-
-  ## Build directory structure
   `%+%` <- function(x, y) paste0(x, y)
+
+  # Make folder structure
   system("mkdir recount2-" %+% ID)
   load(getwd() %+% "/" %+% ID %+% "/rse_gene.Rdata")
   setwd(getwd() %+% "/recount2-" %+% ID)
   system("mkdir 0-make")
   system("mkdir annot")
   system("mkdir data")
-
-  ## Find per-disease label
   gc()
+
+  # Retrieve data from RSE object
   colData <- SummarizedExperiment::colData
   assays <- SummarizedExperiment::assays
   colnames(colData(rse_gene))
   annot <- colData(rse_gene)
-  projects <- do.call("rbind", lapply(annot$characteristics, t))
+
+  # Get V1-VN characteristics
+  projSummary <- function(data){
+    annots <- lapply(data, t)
+    annots <- lapply(annots, as.data.frame)
+    do.call(plyr::rbind.fill, annots)
+  }
+
+  # Ask user which characteristics to use
+  projects <- projSummary(annot$characteristics)
+  if(missing(characteristic)){
+    # Ask for user input
+    lapply(projects, function(x){
+      print(table(x))
+    })
+    characteristic <- readline(paste(
+      "Which characteristic do you want to use to split files?",
+      "[Input Numeric]: "))
+    characteristic <- as.numeric(characteristic)
+  }
+
   projects <- projects[, characteristic]
   project <- table(projects)
-  if(!identical(sum(project), nrow(colData(rse_gene)))) stop()
+  if (!identical(sum(project), nrow(colData(rse_gene))))
+    stop()
   prj <- names(project)
+  for (i in 1:length(prj)) {
 
-  ## Split data by per-disease label
-  for(i in 1:length(prj)){
-
-    # Get counts and annot
+    # Subset the i-th project, scale counts
     rse.i <- rse_gene[, projects == prj[i]]
     rse.i <- recount::scale_counts(rse.i)
     scale_counts <- assays(rse.i)[[1]]
     annot <- colData(rse.i)
-    annot.split <- do.call("rbind", lapply(annot$characteristics, t))
+
+    # Summarize projects for annot.i
+    annot.split <- projSummary(annot$characteristics)
     annot <- cbind(annot, annot.split)
     annot$characteristics <- "See V1-VN"
 
-    # Remove NA columns and commas
+    # Remove all NA annot
     allisNA <- apply(annot, 2, function(x) all(is.na(x)))
     annot <- annot[, !allisNA]
-    for(j in 1:ncol(annot)){
-      if(any(grepl(",", annot[,j]))){
-        annot[,j] <- gsub(",", "-", annot[,j])
+    for (j in 1:ncol(annot)) {
+      if (any(grepl(",", annot[, j]))) {
+        annot[, j] <- gsub(",", "-", annot[, j])
       }
     }
 
-    if(!identical(rownames(annot),
-                  colnames(scale_counts))) stop("")
+    # Check counts.i and annot.i are same
+    if (!identical(rownames(annot), colnames(scale_counts)))
+      stop("")
 
-    # Save counts and annot
+    # Save counts.i and annot.i
     id <- gsub(" ", "_", prj[i])
-    utils::write.csv(scale_counts, file = paste0("data/", id, "-scale_counts.csv"))
+    utils::write.csv(scale_counts, file = paste0("data/",
+                                                 id, "-scale_counts.csv"))
     utils::write.csv(annot, file = paste0("annot/", id, "-annot.csv"))
   }
 
-  # Add miSciTools note
-  utils::write.table("see miSciTools::getRecount", "0-make/0-make.R")
-
+  # Leave miSciTools note
+  utils::write.table(paste("see miSciTools::getRecount v.",
+                           utils::packageVersion("miSciTools")),
+                     file = "0-make/0-make.R")
   setwd(catchwd)
 }